requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Werkzeug==2.2.3
python-dotenv==1.0.1
waitress==3.0.0